Move subscription-related code to a separate file
